requirements.txt
pandas==1.1.2
torchvision==0.7.0
matplotlib==3.3.1
gradio==1.2.3
seaborn==0.11.0
requests==2.24.0
torch==1.6.0
numpy==1.19.1
Pillow==8.0.1
scikit_learn==0.23.2
